LIFT/LIFT/requirements/base.txt
Django==4.2.1
django-extensions==3.2.1
django-ninja==0.21.0
django-ninja-jwt==5.2.5
tqdm==4.65.0